+++ DigitaleBriefwahlTests/WeightedElectionModelTests.cs
@@ -3,6 +3,7 @@
 // (https://opensource.org/licenses/GPL-3.0)
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using DigitaleBriefwahl.Model;
 using NUnit.Framework;
@@ -61,10 +62,122 @@ Kandidat2=Donald Duck
 Kandidat3=Dagobert Duck
 ";
 			var data = ElectionModelTests.ReadIniDataFromString(ini);
-			var model = ElectionModelFactory.Create("Election", data);
+			var sut = ElectionModelFactory.Create("Election", data);
 
 			// Exercise/Verify
-			Assert.That(model.EmptyVotes, Is.EqualTo(new[] { "Mickey Mouse", "Donald Duck"}));
+			Assert.That(sut.EmptyVotes, Is.EqualTo(new[] { "Mickey Mouse", "Donald Duck"}));
+		}
+
+		[Test]
+		public void GetInvalidVotes_AllOk()
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=Weighted
+Stimmen=3
+Kandidat1=Mickey Mouse
+Kandidat2=Donald Duck
+Kandidat3=Dagobert Duck
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var sut = ElectionModelFactory.Create("Election", data);
+
+			var result = sut.GetInvalidVotes(new List<string>(new[]
+				{ "Mickey Mouse", "Donald Duck", "Dagobert Duck"}));
+
+			Assert.That(result.Count, Is.EqualTo(0));
+		}
+
+		[Test]
+		public void GetInvalidVotes_DuplicateName()
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=Weighted
+Stimmen=3
+Kandidat1=Mickey Mouse
+Kandidat2=Donald Duck
+Kandidat3=Dagobert Duck
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var sut = ElectionModelFactory.Create("Election", data);
+
+			var result = sut.GetInvalidVotes(new List<string>(new[]
+				{ "Mickey Mouse", "Donald Duck", "Donald Duck"}));
+
+			Assert.That(result.Count, Is.EqualTo(2));
+			Assert.That(result.First(), Is.EqualTo(1));
+			Assert.That(result.Last(), Is.EqualTo(2));
+		}
+
+		[Test]
+		public void GetInvalidVotes_InvalidName()
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=Weighted
+Stimmen=3
+Kandidat1=Mickey Mouse
+Kandidat2=Donald Duck
+Kandidat3=Dagobert Duck
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var sut = ElectionModelFactory.Create("Election", data);
+
+			var result = sut.GetInvalidVotes(new List<string>(new[]
+				{ "Mickey Mouse", "X", "Donald Duck"}));
+
+			Assert.That(result.Count, Is.EqualTo(1));
+			Assert.That(result.First(), Is.EqualTo(1));
+		}
+
+		[TestCase(new [] { "Mickey Mouse", "Donald Duck"}, 2)]
+		[TestCase(new [] { "Mickey Mouse", "Donald Duck", ""}, 2)]
+		[TestCase(new [] { "", "Mickey Mouse", "Donald Duck", }, 0)]
+		public void GetInvalidVotes_Incomplete(string[] votes, int expectedInvalid)
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=Weighted
+Stimmen=3
+Kandidat1=Mickey Mouse
+Kandidat2=Donald Duck
+Kandidat3=Dagobert Duck
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var sut = ElectionModelFactory.Create("Election", data);
+
+			var result = sut.GetInvalidVotes(new List<string>(votes));
+
+			Assert.That(result.Count, Is.EqualTo(1));
+			Assert.That(result.First(), Is.EqualTo(expectedInvalid));
+		}
+
+		[TestCase(new [] { "Mickey Mouse", "Donald Duck"}, 2)]
+		[TestCase(new [] { "Mickey Mouse", "Donald Duck", ""}, 2)]
+		[TestCase(new [] { "", "Mickey Mouse", "Donald Duck", }, 0)]
+		public void GetInvalidVotes_IncompleteOk(string[] votes, int expectedInvalid)
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=Weighted
+Stimmen=3
+FehlendOk=true
+Kandidat1=Mickey Mouse
+Kandidat2=Donald Duck
+Kandidat3=Dagobert Duck
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var sut = ElectionModelFactory.Create("Election", data);
+
+			var result = sut.GetInvalidVotes(new List<string>(votes));
+
+			Assert.That(result.Count, Is.EqualTo(0));
 		}
 	}
 }